Restructure pyresample cache modules
