+++ pyresample/cache/__init__.py
@@ -17,30 +17,5 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Cache classes and utilities for Resampler classes."""
 
-
-from abc import ABC, abstractmethod
-
-
-class ResampleCache(ABC):
-    """Base class for all BaseResampler cache classes."""
-
-    @abstractmethod
-    def store(self, key, value):
-        ...
-
-    @abstractmethod
-    def load(self, key):
-        ...
-
-
-class InMemoryCache(ResampleCache):
-    """Basic cache that stores everything in-memory."""
-
-    def __init__(self):
-        self._store = {}
-
-    def store(self, key, value):
-        self._store[key] = value
-
-    def load(self, key):
-        return self._store[key]
+from .base import ResampleCache  # noqa
+from .memory import InMemoryCache  # noqa
